+++ R/n_parameters.R
@@ -10,6 +10,9 @@
 #'    term or the instrumental variables be returned? Applies to models
 #'    with zero-inflated and/or dispersion formula, or to models with instrumental
 #'    variable (so called fixed-effects regressions). May be abbreviated.
+#' @param only_estimable Logical, if \code{TRUE}, removes (i.e. does not count)
+#'    non-estimable parameters (which may occur for models with rank-deficient
+#'    model matrix).
 #' @param ... Arguments passed to or from other methods.
 #'
 #' @return The number of parameters in the model.
@@ -35,10 +38,19 @@ n_parameters <- function(x, ...) {
 
 #' @rdname n_parameters
 #' @export
-n_parameters.default <- function(x, ...) {
-  length(unlist(find_parameters(x, effects = "fixed", flatten = FALSE, ...)))
+n_parameters.default <- function(x, only_estimable = FALSE, ...) {
+  params <- get_parameters(x, effects = "fixed", ...)
+  .process_estimable(params, only_estimable)
 }
 
+
+# helper
+.process_estimable <- function(params, only_estimable) {
+  if (isTRUE(only_estimable)) {
+    params <- params[!is.na(params$Estimate), ]
+  }
+  nrow(params)
+}
 
 
 
@@ -46,9 +58,10 @@ n_parameters.default <- function(x, ...) {
 
 #' @rdname n_parameters
 #' @export
-n_parameters.merMod <- function(x, effects = c("fixed", "random"), ...) {
+n_parameters.merMod <- function(x, effects = c("fixed", "random"), only_estimable = FALSE, ...) {
   effects <- match.arg(effects)
-  length(unlist(find_parameters(x, effects = effects, flatten = FALSE, ...)))
+  params <- get_parameters(x, effects = effects, ...)
+  .process_estimable(params, only_estimable)
 }
 
 #' @export
@@ -90,10 +103,12 @@ n_parameters.wbm <- n_parameters.merMod
 n_parameters.MixMod <- function(x,
                                 effects = c("fixed", "random"),
                                 component = c("all", "conditional", "zi", "zero_inflated"),
+                                only_estimable = FALSE,
                                 ...) {
   effects <- match.arg(effects)
   component <- match.arg(component)
-  length(unlist(find_parameters(x, effects = effects, component = component, flatten = FALSE, ...)))
+  params <- get_parameters(x, effects = effects, component = component, ...)
+  .process_estimable(params, only_estimable)
 }
 
 #' @rdname n_parameters
@@ -108,9 +123,13 @@ n_parameters.glmmTMB <- n_parameters.MixMod
 
 #' @rdname n_parameters
 #' @export
-n_parameters.zeroinfl <- function(x, component = c("all", "conditional", "zi", "zero_inflated"), ...) {
+n_parameters.zeroinfl <- function(x,
+                                  component = c("all", "conditional", "zi", "zero_inflated"),
+                                  only_estimable = FALSE,
+                                  ...) {
   component <- match.arg(component)
-  length(unlist(find_parameters(x, component = component, flatten = FALSE, ...)))
+  params <- get_parameters(x, component = component, ...)
+  .process_estimable(params, only_estimable)
 }
 
 #' @export
@@ -128,9 +147,13 @@ n_parameters.zerotrunc <- n_parameters.default
 
 #' @rdname n_parameters
 #' @export
-n_parameters.gam <- function(x, component = c("all", "conditional", "smooth_terms"), ...) {
+n_parameters.gam <- function(x,
+                             component = c("all", "conditional", "smooth_terms"),
+                             only_estimable = FALSE,
+                             ...) {
   component <- match.arg(component)
-  length(unlist(find_parameters(x, component = component, flatten = FALSE, ...)))
+  params <- get_parameters(x, component = component, ...)
+  .process_estimable(params, only_estimable)
 }
 
 #' @export